AR-10 Added a marker that the opcode clean script must not be deleted

+++ src/Opcode/CleanReset.php
@@ -25,9 +25,15 @@ class CleanReset extends AbstractTask
             $this->runtime->runCommand('curl -k -X GET ' . $url . '/apc_clear.php');
         }
 
-        return $this->runtime
-            ->runCommand('rm -rf ' . ($this->options['doc_root'] ?? '.') . '/apc_clear.php')
-            ->isSuccessful();
+        $deleteSuccess = true;
+
+        if (!@$this->options['without_delete']) {
+            $deleteSuccess = $this->runtime
+                ->runCommand('rm -rf ' . ($this->options['doc_root'] ?? '.') . '/apc_clear.php')
+                ->isSuccessful();
+        }
+
+        return $deleteSuccess;
     }
 
     /**
